requirements.txt
plotly==4.6.0
psnparser==1.0
python-shogi=1.0.9
